Refactored dev models. Added neuralpredictors as dev dependency
